Small Updates + Beginning of Hand Evaluation
Changed the cards on the table to not be given magic numbers for spawning, where cards spawn can be changed by changing the Vector3 variables. For some reason the cards can spawn flipped and the cause is currently unknown to me as to how they do that. I also added in some basic hand evaluation in the sense that the cards in hand and on the table are now put into 2 arrays for rank and suit and also a 2d Array for displaying the current deck easier making it easier to check for hand score